requirements.txt
django
django-cors-headers
django-extensions
django-filter
django-heroku
djangorestframework
djangorestframework-simplejwt
django_reverse_admin
djoser
drf-nested-routers
drf-yasg
factory_boy
gunicorn
markdown
python-stdnum
rest_condition
qrcode
